lag hook for behandlingsresultat og flytt opprettelse av endret utbetaling og feilmelding-state inn dit
